requirements.txt
aniso8601==1.3.0
click==6.7
Flask==0.12.2
Flask-JWT-Simple==0.0.3
flask-restplus==0.10.1
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
Werkzeug==0.12.2
